Do int type cast in case rem_cpus has underflowed
*switch_cpu_cnt is uint32_t, so it needs to be cast to int before being
compared to rem_cpus in case rem_cpus has underflowed, which is a valid
state. e.g. when we are getting exclusive node allocations, the job will
reach the CPU requirement (rem_cpus) sooner than it reaches the node
count requirement (rem_nodes).

Note that this is not an issue for switch_node_cnt since it is already
type int.

Ticket 20547

+++ src/plugins/topology/tree/eval_nodes_tree.c
@@ -63,9 +63,9 @@ static int _topo_compare_switches(int i, int j, int rem_nodes,
 {
 	while (1) {
 		bool i_fit = ((switch_node_cnt[i] >= rem_nodes) &&
-			      (switch_cpu_cnt[i] >= rem_cpus));
+			      ((int) switch_cpu_cnt[i] >= rem_cpus));
 		bool j_fit = ((switch_node_cnt[j] >= rem_nodes) &&
-			      (switch_cpu_cnt[j] >= rem_cpus));
+			      ((int) switch_cpu_cnt[j] >= rem_cpus));
 		*i_fit_out = i_fit;
 
 		if (i_fit && j_fit) {
